requirements.txt
langgraph>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
langsmith>=0.1.0
langgraph-cli[inmem]>=0.1.0
langchain>=0.3.0
tavily-python>=0.5.0
tenacity>=8.2.0
pytest